Move ModelV2 to Model
