+++ src/graphdatascience/model/model_details.py
