+++ src/graphdatascience/procedure_surface/arrow/model_api_arrow.py
@@ -5,8 +5,8 @@ from typing import Any
 
 from graphdatascience.arrow_client.authenticated_flight_client import AuthenticatedArrowClient
 from graphdatascience.arrow_client.v2.data_mapper_utils import deserialize
-from graphdatascience.model.v2.model_api import ModelApi
-from graphdatascience.model.v2.model_details import ModelDetails
+from graphdatascience.model.model_api import ModelApi
+from graphdatascience.model.model_details import ModelDetails
 
 
 class ModelApiArrow(ModelApi):
