+++ src/graphdatascience/procedure_surface/cypher/model_api_cypher.py
@@ -3,8 +3,8 @@ from typing import Any
 import neo4j
 
 from graphdatascience.call_parameters import CallParameters
-from graphdatascience.model.v2.model_api import ModelApi
-from graphdatascience.model.v2.model_details import ModelDetails
+from graphdatascience.model.model_api import ModelApi
+from graphdatascience.model.model_details import ModelDetails
 from graphdatascience.query_runner.query_runner import QueryRunner
